#49 refactoring of JsonUtil
